refactor(scripts): file-organization sweep over updateNode
The version and manifest helpers move to services/updateNode with their tests and constants, the registry document shape to models/updateNode, and the entrypoint folder keeps index.ts beside the two platform install scripts the root manifest dispatches to.
