+++ scripts/src/services/updateNode/setDevEnginesRuntime.ts
@@ -1,5 +1,5 @@
-import { DEV_ENGINES_RUNTIME_VERSION_REGEX } from "#src/updateNode/constants";
-import { setVersion } from "#src/updateNode/setVersion";
+import { DEV_ENGINES_RUNTIME_VERSION_REGEX } from "#src/services/updateNode/constants";
+import { setVersion } from "#src/services/updateNode/setVersion";
 
 /** Rewrite a package.json string's `devEngines.runtime.version` to `^${version}`. */
 export const setDevEnginesRuntime = (packageJson: string, version: string): string =>
